Initial commit of Ableton session template and racks
This includes an Ableton Live session file that allows a user to plug in their OP-Z and immiediately control similart instrument tracks, parameters and effects that have been meticulously mapped out and set up in Ableton Live. The purpose of this template is to mimic the instruments, sounds and effects on the OP-Z but within Live.